Integrating the Final API: added function to get token
The code above starts by checking whether an access token was found. If no token is found (!accessToken), it then checks for an authorization code. If no authorization code is found (!code), the user is automatically redirected to the Google Authorization screen, where they can sign in and receive their code